requirements.txt
fastapi
uvicorn
python-multipart
opencv-python
jinja2
pytest
httpx
click
pytest-asyncio
Pillow
python-dotenv
tqdm

